requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
openai>=1.0
caldav==1.3.6
icalendar>=5.0
typer>=0.9.0
python-dotenv==1.0.0
apscheduler>=3.10
cryptography>=41.0
pydantic>=2.0
pytest>=7.0
mypy>=1.0
black>=23.0
pre-commit>=3.0 